+++ apps/vmq_diversity/src/vmq_diversity_script_state.erl
@@ -226,7 +226,8 @@ load_script(Id, Script) ->
             {vmq_diversity_ets,         <<"kv">>},
             {vmq_diversity_lager,       <<"log">>},
             {vmq_diversity_memcached,   <<"memcached">>},
-            {vmq_diversity_cache,       <<"auth_cache">>}
+            {vmq_diversity_cache,       <<"auth_cache">>},
+            {vmq_diversity_vmq_api,     <<"vmq_api">>}
            ],
 
     {ok, ScriptsDir} = application:get_env(vmq_diversity, script_dir),