+++ apps/vmq_diversity/test/vmq_diversity_provider_SUITE.erl
@@ -20,7 +20,8 @@
          bcrypt_test/1,
          logger_test/1,
          memcached_test/1,
-         auth_cache_test/1]).
+         auth_cache_test/1,
+         vmq_api_test/1]).
 
 %% ===================================================================
 %% common_test callbacks
@@ -67,7 +68,8 @@ all() ->
              json_test,
              bcrypt_test,
              logger_test,
-             auth_cache_test]
+             auth_cache_test,
+             vmq_api_test]
     end.
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -117,3 +119,6 @@ memcached_test(_) ->
 
 auth_cache_test(_) ->
     {ok, _} = vmq_diversity:load_script(test_script("cache_test.lua")).
+
+vmq_api_test(_) ->
+    {ok, _} = vmq_diversity:load_script(test_script("vmq_api_test.lua")).